feat: 🎸 connect redux to PersonaQuestion
